requirements.txt
bs4
streamlit
requests_html


